refactor(promotions page): change stage2 according to mobile new design
affects: @haaretz/purchase-page
